back-end-python-django/SIM/requirements.txt
Django==4.2.0
djangorestframework==3.15.1
drf-yasg
django-rest-swagger==2.2.0

# Django==4.2
# djangorestframework
# drf-spectacular

# django-rest-swagger
# drf-yasg
# # Django==3.2.12
# # djangorestframework==3.13.1
# # django-rest-swagger==2.2.0
# # drf-yasg==1.24.0
mysqlclient